requirements.txt
anyio==4.4.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-ai-generativelanguage
google-api-core==2.19.2
google-auth==2.34.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.5
idna==3.10
proto-plus==1.24.0
protobuf
pyasn1==0.6.1
pyasn1-modules==0.4.1
python-dotenv==1.0.1
python-telegram-bot==21.5
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tqdm==4.66.5
urllib3==2.2.3
Pillow==10.4.0
PyPDF2
google-generativeai
PySocks
logging
python-docx
bs4
python-pptx==1.0.2
chardet
pygments
openpyxl
docx2txt